multitenant: split external IO impl out to subpackage
The `cloud` package depends on a bunch of things, the multitenant package did
not other than this one transitive dependency to access the interface in
question.

Relates to #86334.

Release justification: part of larger bug fix, totally mechanical change

Release note: None
