+++ pkg/BUILD.bazel
@@ -1164,6 +1164,7 @@ GO_TARGETS = [
     "//pkg/kv/kvserver:kvserver_test",
     "//pkg/kv:kv",
     "//pkg/kv:kv_test",
+    "//pkg/multitenant/multitenantio:multitenantio",
     "//pkg/multitenant/tenantcostmodel:tenantcostmodel",
     "//pkg/multitenant:multitenant",
     "//pkg/obs:obs",
@@ -2410,6 +2411,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/kv/kvserver/txnwait:get_x_data",
     "//pkg/kv/kvserver/uncertainty:get_x_data",
     "//pkg/multitenant:get_x_data",
+    "//pkg/multitenant/multitenantio:get_x_data",
     "//pkg/multitenant/tenantcostmodel:get_x_data",
     "//pkg/obs:get_x_data",
     "//pkg/obsservice/cmd/obsservice:get_x_data",
